add signed transactions

+++ app/src/main/java/com/cybexmobile/graphene/chain/FeeAmountObject.java
@@ -4,6 +4,6 @@ import java.io.Serializable;
 
 public class FeeAmountObject implements Serializable{
 
-    public double amount;
+    public long amount;
     public String asset_id;
 }
